+++ demos/react/src/components/App.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface AppProps {
+  name: string;
+}
+
+export default function App({ name }: AppProps) {
+  return <section>{name}</section>;
+}